requirements-python.txt
python-dotenv==1.0.0
pandas==2.0.0
rpy2==3.5.10
openai==0.27.2
matplotlib==3.7.1
ptyprocess==0.7.0